feat(prayer-timings): consolidate prayer-timings page modals
Three separate modals on the Prayer Times page were merged down to two:

1. Sunrise/Sunset → folded into the Prayer Time Adjustments modal
   - Delete the standalone Sunrise & Sunset modal and its trigger button.
   - In the adjustments modal, add Sunrise and Sunset accordion items at
     the bottom of the Starts tab only. They share the same
     Default / Offset / Manual controls as the other entries.
   - Save handler writes both prayer adjustments and sunrise/sunset
     adjustments, then bubbles the updated Settings up via onSettingsChange.

2. Calculation + Hijri → merged into one CalculationSettingsModal
   - New tabbed modal: "Prayer" (calculation method, juristic school,
     read-only masjid location) and "Hijri" (current date, hijri
     calculation method, date offset).
   - Save writes both updatePrayerCalculationSettings and
     updateHijriSettings, then notifies the parent through
     onSettingsChange. The page passes forceUpdate via that callback so
     the monthly table refetches when calculation method changes.
   - Drop the standalone Hijri button and modal.
   - Simplify the Hijri "Current Hijri Date" — drop the boxed container,
     show it as an inline muted-label + bold-date row.

3. Bring back the Sunrise column in the monthly prayer-times table
   (between Fajr and Dhuhr). The column applies Settings.sunrise_adjustment
   via applySingleAdjustment and stays constant across the
   Starts/Athan/Iqamah tab toggle (sunrise has no category split).
