+++ src/components/modals/index.ts
@@ -15,14 +15,8 @@ export const MapModal = lazy(() => import('./map-modal').then(m => ({ default: m
 export const PrayerTimingsModal = lazy(() =>
   import('./prayer-timings-modal').then(m => ({ default: m.PrayerTimingsModal }))
 );
-export const PrayerCalculationModal = lazy(() =>
-  import('./prayer-calculation-modal').then(m => ({ default: m.PrayerCalculationModal }))
-);
-export const SunriseSunsetModal = lazy(() =>
-  import('./sunrise-sunset-modal').then(m => ({ default: m.SunriseSunsetModal }))
-);
-export const HijriModal = lazy(() =>
-  import('./hijri-modal').then(m => ({ default: m.HijriModal }))
+export const CalculationSettingsModal = lazy(() =>
+  import('./calculation-settings-modal').then(m => ({ default: m.CalculationSettingsModal }))
 );
 export const ScreenModal = lazy(() =>
   import('./screen-modal').then(m => ({ default: m.ScreenModal }))
